Update AccountsLogic usage and organize CurrentAccount
Update AccountsLogic to use instance methods rather than static in 'UserRegistration.cs' and 'AccountsLogicTests.cs'. Static usage was unnecessary and confused the codebase in instances where it was not required. 'UpdateList' method in 'AccountsLogic.cs' is changed to an instance method for consistency. 'CurrentAccount' property moved from 'AccountsLogic.cs' to 'Helpers.cs' to allow application wide access.
Code is now more modular and easy to understand.